extract utf8ByteLength to shared utility and remove byte array allocations from monitoring support
Move utf8ByteLength into SparkConfUtils so all callers share one
implementation. SparkToonOps and ToonMonitoringSupport both used
getBytes(UTF-8).length to measure string byte sizes which allocates
and immediately discards byte arrays. Now both delegate to the
zero-allocation scanner.
